pydantic_related/pyproject.toml
[project]
name = "pydantic-related"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langfuse>=3.2.1",
    "mypy>=1.17.0",
    "pydantic-ai[examples]>=0.4.3",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
